+++ CHANGES.md
@@ -1,3 +1,27 @@
+# 1.24.0
+
+-   We now have a peer dependency on `decimal.js-light` which defines a `Decimal`
+    object. We support this with a MST type and a converter (read on).
+
+-   We introduced a new `decimal` mobx-state-tree type you can import and use
+    within a MST model:
+
+    ```js
+    const M = types.model({
+        d: decimal
+    });
+    ```
+
+    Its JSON representation is a string, but the property will be a `Decimal`
+    instance.
+
+*   BREAKING CHANGE: the `converters.decimal` converter used to have a string
+    as both its value and its raw. It's been changed to have the `Decimal`
+    object as a value and can thus be used with the `decimal` type. To get the
+    old behavior you need to use `converters.stringDecimal`, so you can convert
+    your decimal converters to this to keep your code working until you change
+    it to use the new `decimal` mobx-state-tree type we introduce.
+
 # 1.23.0
 
 -   BREAKING CHANGE: Change the way sources work: `load` now takes the query as